+++ src/map.rs
@@ -325,11 +325,11 @@ impl Map {
         [
             self.decode_node(CNodeId {
                 tile_number: way_id.tile_number,
-                local_node_id: 2 * way_id.local_way_id,
+                local_node_id: 2 * way_id.local_way_id as u16,
             }),
             self.decode_node(CNodeId {
                 tile_number: way_id.tile_number,
-                local_node_id: 2 * way_id.local_way_id + 1,
+                local_node_id: 2 * way_id.local_way_id as u16 + 1,
             }),
         ]
     }
